fixed some issues with compatibility with new versions of numpy etc

+++ lib/masklib.py
@@ -39,9 +39,9 @@ def load_grid(glacier,xmin,xmax,ymin,ymax,dx,ice=0,icefront_time='none',icefront
   files = os.listdir(DIR)
   
   # Set up grid
-  nx = np.ceil((xmax-xmin)/dx)+1
+  nx = int(np.ceil((xmax-xmin)/dx))+1
   x = np.linspace(xmin,xmin+dx*(nx-1),nx)
-  ny = np.ceil((ymax-ymin)/dx)+1
+  ny = int(np.ceil((ymax-ymin)/dx))+1
   y = np.linspace(ymin,ymin+dx*(ny-1),ny)
   
   if ice == 1:
